+++ tests/test_orm_usage.py
@@ -1,11 +1,11 @@
 """Tests that represent usage as an ORM."""
 
-from dataclasses import dataclass
+from typing import List, Optional
 
 import pytest
 
 from datafiles import datafile
-from datafiles.utils import logbreak
+from datafiles.utils import logbreak, write
 
 
 # This model is based on the example dataclass from:
@@ -48,7 +48,7 @@ def test_multiple_instances_are_distinct(expect):
 
 
 def test_classes_can_share_a_nested_dataclass(expect):
-    @dataclass
+    @datafile
     class Nested:
         value: int
 
@@ -78,3 +78,58 @@ def test_values_are_filled_from_disk(expect):
     items = list(InventoryItem.objects.all())
 
     expect(items[0]) == InventoryItem(42, "Things", 0.99)
+
+
+def test_missing_optional_fields_are_loaded(expect):
+    @datafile
+    class Name:
+        value: str
+
+    @datafile("../tmp/samples/{self.key}.json")
+    class Sample:
+
+        key: int
+        name: Optional[Name]
+        value: float = 0.0
+
+    sample = Sample(42, None)
+
+    logbreak("get key=42")
+    sample2 = Sample.objects.get(42)
+    expect(sample2.name) == sample.name
+
+
+def test_comments_in_matched_files(expect):
+    @datafile("../tmp/templates/{self.key}/config.yml")
+    class LegacyTemplate:
+        key: str
+        name: str
+        link: str
+        default: List[str]
+        aliases: List[str]
+
+    write(
+        'tmp/templates/foo/config.yml',
+        """
+        link: # placeholder
+        default:
+          - # placeholder
+          - # placeholder
+        aliases:
+          - # placeholder
+        """,
+    )
+    write(
+        'tmp/templates/bar/config.yml',
+        """
+        link: http://example.com
+        default:
+          - abc
+          - def
+        aliases:
+          - qux
+        """,
+    )
+
+    items = list(LegacyTemplate.objects.all())
+    expect(len(items)) == 2