+++ tests/test_saving.py
@@ -1,9 +1,12 @@
 """Tests for saving to a file."""
 
-# pylint: disable=unused-variable,assigning-non-slot
+# pylint: disable=unused-variable,assigning-non-slot,unsubscriptable-object
+
+from typing import Optional
 
 import pytest
 
+from datafiles import datafile
 from datafiles.utils import dedent, logbreak, read, write
 
 from .samples import (
@@ -224,6 +227,26 @@ def describe_optionals():
             """
         )
 
+    def when_nested_dataclass_is_none(expect):
+        @datafile
+        class Name:
+            value: str
+
+        @datafile("../tmp/samples/{self.key}.yml")
+        class Sample:
+
+            key: int
+            name: Optional[Name]
+            value: float = 0.0
+
+        sample = Sample(42, None)
+
+        expect(read('tmp/samples/42.yml')) == dedent(
+            """
+            name:
+            """
+        )
+
 
 def describe_defaults():
     def with_custom_values(expect):
@@ -293,8 +316,8 @@ def describe_preservation():
         write(
             'tmp/sample.yml',
             """
-            # Heading
-            required: 1.0  # Line
+            # Header
+            required: 1.0       # Line
             optional: 2.0
             """,
         )
@@ -305,8 +328,8 @@ def describe_preservation():
 
         expect(read('tmp/sample.yml')) == dedent(
             """
-            # Heading
-            required: 3.0  # Line
+            # Header
+            required: 3.0       # Line
             optional: 2.0
             """
         )
@@ -317,76 +340,112 @@ def describe_preservation():
         write(
             'tmp/sample.yml',
             """
-            # Heading
+            # Header
             name: a
-            score: 1.0  # Line
+            score: 1.0      # Line
 
             nested:
-              # Nested heading
+              # Nested header
               name: n
               score: 2
             """,
         )
 
-        logbreak("Loading")
         sample.datafile.load()
-
         sample.score = 3
         sample.nested.score = 4
-
-        logbreak("Saving")
         sample.datafile.save()
 
         expect(read('tmp/sample.yml')) == dedent(
             """
-            # Heading
+            # Header
             name: a
-            score: 3.0  # Line
+            score: 3.0      # Line
 
             nested:
-              # Nested heading
+              # Nested header
               name: n
               score: 4.0
             """
         )
 
-    @pytest.mark.xfail(reason="Unknown ruamel.yaml bug")
     def with_comments_on_nested_lines(expect):
         sample = SampleWithNestingAndDefaults(None)
 
         write(
             'tmp/sample.yml',
             """
-            # Heading
+            # Header
             name: a
-            score: 1  # Line
+            score: 1        # Line
 
             nested:
-              # Nested heading
+              # Nested header
               name: n
-              score: 2  # Nested line
+              score: 2      # Nested line
             """,
         )
 
-        logbreak("Loading")
         sample.datafile.load()
-
         sample.score = 3
         sample.nested.score = 4
-
-        logbreak("Saving")
         sample.datafile.save()
 
         expect(read('tmp/sample.yml')) == dedent(
             """
-            # Heading
+            # Header
             name: a
-            score: 3.0  # Line
+            score: 3.0      # Line
 
             nested:
-              # Nested heading
+              # Nested header
               name: n
-              score: 4.0  # Nested line
+              score: 4.0    # Nested line
+            """
+        )
+
+    @pytest.mark.xfail(reason="https://github.com/jacebrowning/datafiles/issues/171")
+    def with_comments_on_list_items(expect):
+        sample = SampleWithListOfDataclasses()
+
+        write(
+            'tmp/sample.yml',
+            """
+            # Header
+
+            items:  # Subheader
+
+              # Section
+
+              - name: a     # Item
+                score: 1
+
+              # Section
+
+              - name: b     # Item
+                score: 2
+            """,
+        )
+
+        sample.datafile.load()
+        sample.items[1].name = 'c'
+        sample.datafile.save()
+
+        expect(read('tmp/sample.yml')) == dedent(
+            """
+            # Header
+
+            items:  # Subheader
+
+              # Section
+
+              - name: a     # Item
+                score: 1
+
+              # Section
+
+              - name: c     # Item
+                score: 2
             """
         )
 
@@ -401,6 +460,7 @@ def describe_preservation():
         )
 
         sample.datafile.load()
+        sample.float_ = 1
         sample.datafile.save()
 
         expect(read('tmp/sample.yml')) == dedent(
@@ -408,6 +468,6 @@ def describe_preservation():
             str_: "42"
             bool_: false
             int_: 0
-            float_: 0.0
+            float_: 1.0
             """
         )